Split webview hooks
