+++ src/actions.js
@@ -4,6 +4,7 @@ export const ABOUT_DIALOG_TOGGLE_UPDATE_ON_START = 'about-dialog/toggle-update-o
 export const ADD_SERVER_VIEW_SERVER_ADDED = 'add-server/view-server-added';
 export const CERTIFICATE_TRUST_REQUESTED = 'certificate/trust-requested';
 export const CERTIFICATES_CHANGED = 'certificates/changed';
+export const CERTIFICATES_CLEARED = 'certificates/cleared';
 export const DEEP_LINK_TRIGGERED = 'deep-link/triggered';
 export const LOADING_ERROR_VIEW_RELOAD_SERVER_CLICKED = 'loading-error-view/reload-server-clicked';
 export const MAIN_WINDOW_STATE_CHANGED = 'main-window/state-changed';
@@ -65,8 +66,6 @@ export const WEBVIEW_LOADING_DONE = 'webview/loading-done';
 export const WEBVIEW_LOADING_FAILED = 'webview/loading-failed';
 export const WEBVIEW_LOADING_STARTED = 'webview/loading-started';
 export const WEBVIEW_SCREEN_SHARING_SOURCE_REQUESTED = 'webview/screen-sharing-source-requested';
-export const WEBVIEW_SHORTCUT_KEY_DOWN = 'webview/shortcut-key-down';
-export const WEBVIEW_SHORTCUT_KEY_UP = 'webview/shortcut-key-up';
 export const WEBVIEW_SIDEBAR_STYLE_CHANGED = 'webview/sidebar-style-changed';
 export const WEBVIEW_TITLE_CHANGED = 'webview/title-changed';
 export const WEBVIEW_UNREAD_CHANGED = 'webview/unread-changed';
